test(button): add unit test
